templates/sentiment/requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
mlflow>=2.9.0
PyYAML>=6.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0

# API serving dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
